# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tekstil Otomasyonu API - OpenAPI 3.0",
        "license": {
            "name": "MIT License",
            "url": "https://mit-license.org/"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:5000"
        }
    ],
...
    "paths": {
...
        "/urunler": {
            "post": {
                "tags": [
                    "urunler"
                ],
                "summary": "Ürün oluşturur.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UrunWithoutID"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "İşlem başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Urun"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ürün bilgileri eksik."
                    },
                    "500": {
                        "description": "Ürün Kaydedilemedi."
                    }
                }
            },
            "get": {
                "tags": [
                    "urunler"
                ],
                "summary": "Ürünleri Getirir",
                "responses": {
                    "200": {
                        "description": "Ürünler Getirildi.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Urun"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ürünler bulunamadı."
                    }
                }
            }
        },
        "/urunler/{id}": {
            "get": {
                "tags": [
                    "urunler"
                ],
                "summary": "ID bilgisine göre ürün getirir.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "İstenen ürünün ID bilgisi",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "İstenen ürün getirildi.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Urun"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ürün bulunamadı."
                    }
                }
            },
            "put": {
                "tags": [
                    "urunler"
                ],
                "summary": "Ürün günceller.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Güncellenecek ürünün ID bilgisi",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UrunWithoutID"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Ürün Güncellendi.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Urun"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ürün Bulunamadı."
                    },
                    "500": {
                        "description": "Ürün Güncellenemedi."
                    }
                }
            },
            "delete": {
                "tags": [
                    "urunler"
                ],
                "summary": "Ürün siler.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Silinmek istenen ürünün ID bilgisi",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "İstenen ürün silindi.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Urun"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ürün bulunamadı."
                    },
                    "500": {
                        "description": "Ürün silinemedi."
                    }
                }
            },
            "patch": {
                "tags": [
                    "urunler"
                ],
                "summary": "Ürün günceller.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Güncellenecek ürünün ID bilgisi",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UrunWithoutID"
                            }
                        }
                    },
                    "required": false
                },
                "responses": {
                    "200": {
                        "description": "İşlem başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Urun"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ürün bilgileri eksik."
                    },
                    "500": {
                        "description": "Ürün Kaydedilemedi."
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Urun": {
                "required": [
                    "urun_id",
                    "urun_adi",
                    "paket_miktar",
                    "paket_fiyat",
                    "stok"
                ],
                "type": "object",
                "properties": {
                    "urun_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 14
                    },
                    "urun_adi": {
                        "type": "string",
                        "example": "Zürafalı Kalem"
                    },
                    "paket_miktar": {
                        "type": "integer",
                        "format": "int64",
                        "example": 5
                    },
                    "paket_fiyat": {
                        "type": "number",
                        "example": 50
                    },
                    "stok": {
                        "type": "integer",
                        "format": "int64",
                        "example": 237
                    }
                }
            },
            "UrunWithoutID": {
                "required": [
                    "urun_adi",
                    "paket_miktar",
                    "paket_fiyat",
                    "stok"
                ],
                "type": "object",
                "properties": {
                    "urun_adi": {
                        "type": "string",
                        "example": "Zürafalı Kalem"
                    },
                    "paket_miktar": {
                        "type": "integer",
                        "format": "int64",
                        "example": 5
                    },
                    "paket_fiyat": {
                        "type": "number",
                        "example": 50
                    },
                    "stok": {
                        "type": "integer",
                        "format": "int64",
                        "example": 237
                    }
                }
            }
        }
    }
}